restructure to match my mental model of ruby code
