+++ mars-rover/.rspec
@@ -1,0 +1,3 @@
+--color
+--require spec_helper
+-f d
